+++ packages/core/src/index.ts
@@ -1,0 +1,7 @@
+export { Primitive } from './Primitive';
+export type {
+  PrimitiveProps,
+  PolymorphicComponentProps,
+  PolymorphicComponent,
+} from './Primitive';
+export type { CommonProps } from './types';