+++ packages/forma-36-react-components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.84.2](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-components@3.84.1...@contentful/forma-36-react-components@3.84.2) (2021-03-15)
+
+**Note:** Version bump only for package @contentful/forma-36-react-components
+
+## [3.84.1](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-components@3.84.0...@contentful/forma-36-react-components@3.84.1) (2021-03-15)
+
+**Note:** Version bump only for package @contentful/forma-36-react-components
+
 # [3.84.0](https://github.com/contentful/forma-36/compare/@contentful/forma-36-react-components@3.83.1...@contentful/forma-36-react-components@3.84.0) (2021-03-12)
 
 =======